added moment and updated readme
