+++ README.md
@@ -1,5 +1,15 @@
 # Farmony
 
+
+## To Run
+
+Create a `.env`
+Add the following variables
+`REACT_APP_SERVER_URL={backend url}`
+`CLOUDINARY_URL=https://api.cloudinary.com/v1_1/ddmbb2ian/image/upload`
+
+Be warned, whiile I work on a better alternative, currently the above cloudinary URL is my personal account. I will able to see uploaded photos on the farmony app. You can sign up to cloudinary for free and use your personal account.
+
 ## Created by:
 Mocha Brown
 Nikki Salehi
