+++ package.json
@@ -26,6 +26,7 @@
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "jwt-decode": "^3.1.2",
+    "moment-timezone": "^0.5.41",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-moment": "^1.1.3",
